% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colsets.R
\name{active_acc_colsets}
\alias{active_acc_colsets}
\title{Identify acceleration columns present in a \code{move2} object}
\usage{
active_acc_colsets(x)
}
\arguments{
\item{x}{A \code{move2} containing acceleration data as collected by EOBS,
Ornitela, or similar tracking devices. Most of the time this will be
either loaded from disk using \link[move2:mt_read]{move2::mt_read} or downloaded using
\link[move2:movebank_download_study]{move2::movebank_download_study}.}
}
\value{
A list of character vectors containing names of acceleration column
sets
}
\description{
These functions identify the available columns of acceleration data that can
be used when constructing an \code{acc} vector with \code{as_acc()}.

\code{active_acc_colsets()} determines the columns that will be used by default to
construct acceleration bursts with \code{as_acc()}. For information about how
default columns are selected, see the details below.
}
\details{
By default, if only one full set of acceleration columns is detected, it is
considered the default. Note that some column sets
(\code{\link[=acc_eobs_cols]{acc_eobs_cols()}}, \code{\link[=acc_burst_cols]{acc_burst_cols()}}) must be present in their entirety,
while all others can include subsets of the full set.

If multiple column sets are detected, the set that contains data values is
used as the default.

If multiple column sets are present and contain data values, sets will be
selected in the order returned by \code{\link[=valid_acc_colsets]{valid_acc_colsets()}}.

It is not possible to reset the default column sets for a given \code{move2}
without modifying its columns. However, you can manually provide a column
set to the \code{acc_cols} argument of \code{\link[=as_acc]{as_acc()}} to construct bursts from
non-default columns.
}
\examples{
active_acc_colsets(albatrosses())

# Multiple colsets may be available
active_acc_colsets(move2::mt_stack(albatrosses(), gulls()))

# Missing columns are not included in the set
g <- gulls()
g$acceleration_raw_x <- NULL
active_acc_colsets(g)

# Columns with no data are also removed
g$acceleration_raw_y <- NA
active_acc_colsets(g)

# Some column sets must be present in their entirety
alb <- albatrosses()
alb$eobs_acceleration_axes <- NULL

\dontrun{
  active_acc_colsets(alb)
}
}
