% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acc_transform.R
\name{acc_calibrate}
\alias{acc_calibrate}
\title{Transform raw acceleration values to physical units in an \code{acc} vector}
\usage{
acc_calibrate(acc, calibration)
}
\arguments{
\item{acc}{An \code{acc} vector.}

\item{calibration}{Transformation function to apply to each burst in \code{acc}. See
\code{\link[=acc_calibration]{acc_calibration()}} to specify per-element transformation functions.}
}
\value{
An \code{acc} vector of the same length as the input with transformed
burst matrices.
}
\description{
Applies a burst-level transformation to every burst in an \code{acc} vector.
Specify a transformation with \code{\link[=acc_calibration]{acc_calibration()}}.
}
\examples{
a <- acc_example()

acc_calibrate(a, acc_calibration("ornitela"))

acc_calibrate(a, acc_calibration("eobs", tag_id = 1000))

acc_calibrate(a, acc_calibration(offset = 2048, slope = 0.001))

acc_calibrate(a, acc_calibration(offset = c(2048, 2046), slope = 0.001))
}
\seealso{
\code{\link[=acc_calibration]{acc_calibration()}} and \code{\link[=as_acc_calibration]{as_acc_calibration()}} to
construct transformation functions for use with \code{acc_calibrate()}.
}
