% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colsets.R
\name{acc_colset}
\alias{acc_colset}
\title{Specify acceleration data columns}
\usage{
acc_colset(
  x = NULL,
  y = NULL,
  z = NULL,
  bursts = NULL,
  axes = NULL,
  frequency = NULL
)
}
\arguments{
\item{x, y, z}{In long-format data, the column name(s) for the
X, Y, and/or Z acceleration axes.}

\item{bursts}{For burst-format data, the column name containing the raw burst
strings.}

\item{axes}{For burst-format data, the column name containing the axis
labels.}

\item{frequency}{For burst-format data, the column name containing the
sampling frequency.}
}
\value{
An \code{acc_colset} object.
}
\description{
Define which columns in a \code{move2} object contain acceleration data.
Use this function to manually specify the columns that should be used when
constructing an \code{acc} vector with \code{as_acc()}.

For long-format data (one measurement per row), specify axis columns
with \code{x}, \code{y}, and/or \code{z}.

For burst-format data (where each row contains burst data in a single string),
specify the columns containing the raw burst data, axes, and sampling
frequency.
}
\examples{
# Specify the column names for long-format data for each axis
acc_colset(x = "my_acc_x", y = "my_acc_y", z = "my_acc_z")

# Long format data may consist of a subset of axes
acc_colset(x = "my_acc_x", y = "my_acc_y")

# Specify the column names for the bursts, axes, and frequency for
# burst-format data
acc_colset(
  bursts = "my_raw_acc",
  axes = "my_axes",
  frequency = "my_freq"
)
}
\seealso{
\code{\link[=active_acc_colsets]{active_acc_colsets()}} for automatic colset detection, \code{\link[=as_acc]{as_acc()}} to
construct an \code{acc} vector using a colset specification.
}
