% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acc_set_units.R
\name{acc_set_units}
\alias{acc_set_units}
\title{Set or convert units on burst data in an \code{acc} vector}
\usage{
acc_set_units(x, units)
}
\arguments{
\item{x}{An \code{acc} vector.}

\item{units}{Character specifying the target units (e.g., \code{"m/s^2"}). For
units in terms of gravitational acceleration, use \code{"standard_free_fall"}.}
}
\value{
An \code{acc} vector with units attached to its burst matrices.
}
\description{
Attach new units or convert existing units to all bursts in an input \code{acc}
vector. See \code{\link[units:units]{units::set_units()}}.

Note that this function simply attaches or converts directly between unit
definitions. To calibrate the raw accelerometer measurements, use
\code{\link[=acc_calibrate]{acc_calibrate()}}.
}
\examples{
a <- acc_example()

# Attach units to unitless bursts
acc_set_units(a, "m/s^2")

# Convert between units
a_ms2 <- acc_set_units(a, "m/s^2")
acc_set_units(a_ms2, "standard_free_fall")
}
\seealso{
\code{\link[=acc_calibrate]{acc_calibrate()}} to calibrate raw acceleration values.
}
