% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acc_transform.R
\name{acc_transform}
\alias{acc_transform}
\title{Transform raw acceleration values to physical units in an \code{acc} vector}
\usage{
acc_transform(acc, .f)
}
\arguments{
\item{acc}{An \code{acc} vector.}

\item{.f}{Transformation function to apply to each burst in \code{acc}. See
\code{\link[=acc_transformer]{acc_transformer()}} to specify per-element transformation functions.}
}
\value{
An \code{acc} vector of the same length as the input with transformed
burst matrices.
}
\description{
Applies a burst-level transformation to every burst in an \code{acc} vector.
Specify a transformation with \code{\link[=acc_transformer]{acc_transformer()}}.
}
\examples{
a <- acc_example()

acc_transform(a, acc_transformer("ornitela"))

acc_transform(a, acc_transformer("eobs", tag_id = 1000))

acc_transform(a, acc_transformer(offset = 2048, slope = 0.001))

acc_transform(a, acc_transformer(offset = c(2048, 2046), slope = 0.001))
}
\seealso{
\code{\link[=acc_transformer]{acc_transformer()}} and \code{\link[=as_acc_transformer]{as_acc_transformer()}} to
construct transformation functions for use with \code{acc_transform()}.
}
