% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acc_transform.R
\name{acc_transformer}
\alias{acc_transformer}
\alias{as_acc_transformer}
\title{Create a list of transformation functions for raw acceleration values}
\usage{
acc_transformer(
  manufacturer = NULL,
  tag_id = NULL,
  sensitivity = NULL,
  offset = NULL,
  offset_x = offset,
  offset_y = offset,
  offset_z = offset,
  slope = NULL,
  slope_x = slope,
  slope_y = slope,
  slope_z = slope,
  orientation = NULL,
  orientation_x = orientation,
  orientation_y = orientation,
  orientation_z = orientation,
  units = "m/s^2",
  axes = "XYZ"
)

as_acc_transformer(df)
}
\arguments{
\item{manufacturer}{Manufacturer of the tag. Currently, \code{"eobs"} and
\code{"ornitela"} are supported. For other manufacturers, leave \code{NULL} and
manually specify the transformation parameters below.}

\item{tag_id}{If \code{manufacturer = "eobs"}, the e-obs tag ID for the tag.}

\item{sensitivity}{If \code{manufacturer = "eobs"}, the sensitivity of the tag.
Defaults to \code{"low"} if none provided. Note that only e-obs generation 1
tags have a sensitivity setting.}

\item{offset, offset_x, offset_y, offset_z}{Custom offset to use in the
transformation. To specify axis-specific offsets, use \code{offset_x},
\code{offset_y}, and/or \code{offset_z}.

Required if no \code{manufacturer} is specified.}

\item{slope, slope_x, slope_y, slope_z}{Custom slope to use in the
transformation. To specify axis-specific slope, use \code{slope_x},
\code{slope_y}, and/or \code{slope_z}.

Required if no \code{manufacturer} is specified.}

\item{orientation, orientation_x, orientation_y, orientation_z}{Either \code{1} or
\code{-1} indicating the orientation of the tag's axes. To
specify axis-specific orientations, use \code{orientation_x}, \code{orientation_y},
and/or \code{orientation_z}. Defaults to \code{1}.

This is useful to standardize orientations across tags of different
manufacturers or generations.}

\item{units}{Output units. Either \code{"m/s^2"} (default) or \code{"standard_free_fall"}.}

\item{axes}{Character string specifying which axes to transform, e.g.,
\code{"XYZ"} (default), \code{"XY"}, or \code{"Z"}. Only these axes will appear in the
transformed output.}

\item{df}{data.frame containing columns corresponding to the available
arguments in \code{acc_transformer()}}
}
\value{
A list of transformation functions.
}
\description{
Generate a list of functions with various transformation parameters to be
used in \code{\link[=acc_transform]{acc_transform()}}.

Use \code{acc_transformer()} to specify transformation parameters manually.
Arguments are vectorized and matched by index.

Use \code{as_acc_transformer()} to convert a data.frame containing rowwise
burst transformation parameters to a list of transformation functions.

This allows you to specify burst-specific transformation functions to
flexibly convert raw acceleration values to physical units in \code{acc} vectors
that contain data from heterogeneous sources.
}
\examples{
# Transformation function for ornitela tags:
acc_transformer(manufacturer = "ornitela")

# E-obs tag defaults vary by tag_id and sensitivity (default `"low"`)
acc_transformer(manufacturer = "eobs", tag_id = 1000, sensitivity = "high")
acc_transformer(manufacturer = "eobs", tag_id = 4000)

# Provide vector arguments to generate elementwise transformation functions:
acc_transformer(
  manufacturer = c("eobs", "ornitela"),
  tag_id = c(1000, NA)
)

# Transformation with explicit offset and slope
acc_transformer(offset = 2048, slope = 1 / 512)

# Transform specific axes with axis-specific args:
tfrm <- acc_transformer(
  offset_x = 2048, 
  offset_y = 2046,
  offset_z = 2048,
  slope = 1 / 512, 
  orientation_y = -1 # Flip y axis orientation
)

# Apply transformations with acc_transform()
acc_transform(acc_example(), tfrm)
}
\seealso{
\code{\link[=acc_transform]{acc_transform()}} to apply transformation functions to the entries
in an \code{acc} vector.
}
