% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colsets.R
\name{active_acc_cols}
\alias{active_acc_cols}
\title{Get acceleration columns from a \code{move2} object}
\usage{
active_acc_cols(x, quiet = FALSE)
}
\arguments{
\item{x}{A \code{move2} containing acceleration data as collected by EOBS,
Ornitela, or similar tracking devices. Most of the time this will be
either loaded from disk using \link[move2:mt_read]{move2::mt_read} or downloaded using
\link[move2:movebank_download_study]{move2::movebank_download_study}.}

\item{quiet}{Logical indicating whether to warn if multiple acceleration
column sets are detected in \code{x}.}
}
\value{
Character vector of acceleration column names
}
\description{
Get the names of the acceleration columns that will be used when parsing
acceleration bursts in a \code{move2} object using \code{as_acc()}.

In the event
that multiple acceleration column sets are present in the input object,
the set that contains data values will be identified as the active
column set. If multiple sets contain data values, the first such set will
be used.
}
