% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colsets.R
\name{active_colsets}
\alias{active_colsets}
\alias{active_acc_colsets}
\alias{active_mag_colsets}
\alias{active_gyro_colsets}
\title{Identify IMU columns present in a \code{move2} object}
\usage{
active_acc_colsets(x)

active_mag_colsets(x)

active_gyro_colsets(x)
}
\arguments{
\item{x}{A \code{move2} object.}
}
\value{
A list of \code{imu_colset} objects.
}
\description{
Determine the column sets that will be used by default when extracting IMU
data from a \code{move2} object. Column sets are processed independently, but a
single \code{move2} may contain multiple active column sets for one IMU sensor.
\itemize{
\item \code{active_acc_colsets()} — column sets used by \code{\link[=as_acc]{as_acc()}}.
\item \code{active_mag_colsets()} — column sets used by \code{\link[=as_mag]{as_mag()}}.
\item \code{active_gyro_colsets()} — column sets used by \code{\link[=as_gyro]{as_gyro()}}.
}
}
\details{
If no active colsets are found, use \code{\link[=imu_colset]{imu_colset()}} to specify the columns
that contain IMU data.
}
\examples{
active_acc_colsets(albatrosses())

# Multiple colsets may be available
active_acc_colsets(move2::mt_stack(albatrosses(), gulls()))

# Missing long-format axes are not included in the set
g <- gulls()
g$acceleration_raw_x <- NULL
active_acc_colsets(g)

# Columns with no data are also removed
g$acceleration_raw_y <- NA
active_acc_colsets(g)

# Some column sets must be present in their entirety
alb <- albatrosses()
alb$eobs_acceleration_axes <- NULL

\dontrun{
  active_acc_colsets(alb)
}
}
\seealso{
\code{\link[=movebank_acc_colsets]{movebank_acc_colsets()}}, \code{\link[=movebank_mag_colsets]{movebank_mag_colsets()}},
\code{\link[=movebank_gyro_colsets]{movebank_gyro_colsets()}} for the supported default colsets.

\code{\link[=as_acc]{as_acc()}}, \code{\link[=as_mag]{as_mag()}}, \code{\link[=as_gyro]{as_gyro()}} to extract IMU data from a
\code{move2} object.
}
