% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colsets.R
\name{active_mag_colsets}
\alias{active_mag_colsets}
\title{Identify magnetometer columns present in a \code{move2} object}
\usage{
active_mag_colsets(x)
}
\arguments{
\item{x}{A \code{move2} containing magnetometer data. Most of the time this will be
either loaded from disk using \link[move2:mt_read]{move2::mt_read} or downloaded using
\link[move2:movebank_download_study]{move2::movebank_download_study}.}
}
\value{
A list of \code{mag_colset} objects.
}
\description{
\code{active_mag_colsets()} determines the sets of columns that will be used
by default to construct magnetometer bursts with \code{as_mag()}. Column sets
are processed independently, but a given \code{move2} may contain multiple
column sets with magnetometer data.
}
\details{
\code{move2} objects store magnetometer data in two ways: long-format
magnetometer columns and burst-format magnetometer columns.

Long-format columns store one magnetometer measurement (possibly for multiple
axes) in a single row. Note that
not all axes need to be present for a long-format column set to be considered
active in a \code{move2} object.

Burst-format columns store a burst of magnetometer data as a space-delimited
string. This string must be segmented into axis-specific measurements using
an associated column that indicates the axes present for the bursted data.
A further column provides the sampling frequency of the burst.
All three of these columns must be present for a burst-format column set
to be considered active in a \code{move2} object.

Standard long-format column sets for data from
Movebank include \code{\link[=mag_colset_raw_xyz]{mag_colset_raw_xyz()}} and \code{\link[=mag_colset_xyz]{mag_colset_xyz()}}. The standard
burst-format column set for data from Movebank is \code{\link[=mag_colset_burst]{mag_colset_burst()}}.

If your input data use different column names for these columns, use
\code{\link[=mag_colset]{mag_colset()}} to specify the column names that correspond to each of the
axes (for long-format data) or burst data and associated metadata (for
burst-format data).
}
\seealso{
\code{\link[=valid_mag_colsets]{valid_mag_colsets()}} for currently supported default colsets,
\code{\link[=as_mag]{as_mag()}} to build a \code{mag} vector from a \code{move2} object.
}
