% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{albatrosses}
\alias{albatrosses}
\title{Galapagos albatrosses}
\format{
A \code{move2} with 9 tracks and 54 features lasting from 2008-07-27 00:00:00 UTC
to 2008-07-27 01:00:00 UTC.

See the \href{https://www.movebank.org/cms/movebank-content/movebank-attribute-dictionary}{Movebank Attribute Dictionary}
for details on data attributes.
}
\source{
\url{https://www.movebank.org/cms/webapp?gwt_fragment=page=studies,path=study2911040}
}
\usage{
albatrosses
}
\description{
GPS and acceleration data for Galapagos albatrosses from Movebank study
2911040. Waved albatrosses were tracked during breeding and non-breeding
periods between 2008 and 2010.
}
\keyword{datasets}
