% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_gyro.R
\name{as_gyro}
\alias{as_gyro}
\alias{as_gyro.default}
\alias{as_gyro.move2}
\title{Convert to gyro}
\usage{
as_gyro(x, ...)

\method{as_gyro}{default}(x, ...)

\method{as_gyro}{move2}(
  x,
  colset = NULL,
  min_freq = 1,
  merge_continuous = TRUE,
  drop = FALSE,
  ...
)
}
\arguments{
\item{x}{A \code{move2} containing gyroscope data. Most of the time this will be
either loaded from disk using \link[move2:mt_read]{move2::mt_read} or downloaded using
\link[move2:movebank_download_study]{move2::movebank_download_study}.}

\item{...}{currently not used}

\item{colset}{An \code{imu_colset} object or list of \code{imu_colset} objects
specifying the columns of \code{x} that contain gyroscope data. By default,
constructs bursts for all column sets that are detected in \code{x} that also
contain data (see \code{\link[=active_gyro_colsets]{active_gyro_colsets()}}).

Several common colsets are listed under \code{\link[=movebank_gyro_colsets]{movebank_gyro_colsets()}}. To
specify a custom set of columns, use \code{\link[=imu_colset]{imu_colset()}}.}

\item{min_freq}{Numeric value indicating the
minimum allowable within-burst data collection frequency when identifying
bursts in long-format data. Any two adjacent timestamps
that fall outside of the period defined by this frequency will be split
into separate bursts. If no units are provided, this value is assumed to
be in Hz.

Ignored if data are already in predefined bursts.}

\item{merge_continuous}{Logical value indicating whether to merge
adjacent bursts. Two adjacent bursts can be merged if the
first burst ends at the same time that the second starts and the
burst frequency is identical between the two. This is useful for
processing continuous data that have been stored in chunks
split at regular intervals.}

\item{drop}{Logical indicating whether empty bursts should
be dropped from the output. If \code{drop = FALSE}, then the length of the
output will match the number of rows in the input data \code{x} and
bursts will be stored at the index location corresponding to the start time
of the burst.}
}
\description{
In many cases the \code{as_gyro} function will directly create a gyroscope
vector from input data
}
\details{
The resulting vector will be as long as the input. This means it
can, for example, be added as a column to a \code{data.frame}. For some tags
this means \code{NA} values are inserted when one burst is stored over multiple
rows of a \code{data.frame}.
}
