% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colsets.R
\name{duplicated_gyro_rows}
\alias{duplicated_gyro_rows}
\title{Identify rows with gyroscope data from multiple column sets}
\usage{
duplicated_gyro_rows(x, colsets = NULL)
}
\arguments{
\item{x}{A \code{move2} containing gyroscope data. Most of the time this will be
either loaded from disk using \link[move2:mt_read]{move2::mt_read} or downloaded using
\link[move2:movebank_download_study]{move2::movebank_download_study}.}

\item{colsets}{List of \code{imu_colset} objects to check for
overlap. Defaults to the column sets detected by \code{\link[=active_gyro_colsets]{active_gyro_colsets()}}.}
}
\value{
An integer vector of row indices with duplicated gyroscope data
across column sets.
}
\description{
This function returns the row indices of a \code{move2} object
where more than one gyroscope column set contains data. \code{as_gyro()}
refuses to build \code{gyro} objects for rows where multiple input sources exist.
These rows can be modified to remove data from additional column sets.
Alternatively, specific columns can be passed to the \code{colset} argument
of \code{as_gyro()} to avoid processing duplicated records.
}
\seealso{
\itemize{
\item \code{\link[=active_gyro_colsets]{active_gyro_colsets()}} to identify available column sets in a \code{move2} object.
\item \code{\link[=as_gyro]{as_gyro()}} to generate a \code{gyro} vector from a \code{move2} object.
}
}
