% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{n_axis}
\alias{n_axis}
\alias{explore-functions}
\alias{n_samples}
\alias{is_uniform}
\alias{is.na.acc}
\alias{bursts}
\alias{freqs}
\alias{starts}
\alias{burst_dur}
\alias{burst_n}
\alias{is_acc}
\title{Functions to explore an \code{acc} vector}
\usage{
n_axis(x)

n_samples(x)

is_uniform(x)

\method{is.na}{acc}(x)

bursts(x)

freqs(x)

starts(x)

burst_dur(x)

burst_n(x)

is_acc(x)
}
\arguments{
\item{x}{an acc vector}
}
\description{
Functions to explore an \code{acc} vector
}
\examples{
x <- acc(
  bursts = list(
    cbind(x = sin(1:30 / 10), y = cos(1:30 / 10), z = 1),
    cbind(x = sin(1:20 / 10 + 2), y = cos(1:20 / 10 + 3))
  ),
  frequency = units::as_units(c(20, 30), "Hz")
)
x <- c(x, NA)
n_axis(x)
n_samples(x)
is_uniform(x)
length(x)
is.na(x)
na.omit(x)
 y <- acc(
  bursts = list(
    cbind(x = sin(1:20 / 10), y = cos(1:20 / 10)),
    cbind(x = sin(1:20 / 10 + 2), y = cos(1:20 / 10 + 3))
  ),
  frequency = units::as_units(c(20, 20), "Hz")
)
is_uniform(y)
}
