% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{filter_freq}
\alias{filter_freq}
\title{Filter an IMU vector by burst frequency}
\usage{
filter_freq(x, min_freq = 0, max_freq = Inf, keep_na = FALSE)
}
\arguments{
\item{x}{An IMU vector (e.g. \code{acc}, \code{mag}, \code{gyro}).}

\item{min_freq, max_freq}{Numeric or units values indicating the minimum
and/or maximum frequency thresholds to use when determining the records in
\code{x} to retain. Elements in \code{x} whose frequency falls within these limits
are kept in the output. If no units are provided, values are considered to
be in hertz.}

\item{keep_na}{Logical indicating whether elements of \code{x} with a missing
frequency should be retained in the output. By default, these elements
are removed.}
}
\value{
A vector of the same class as \code{x}.
}
\description{
Filter an IMU vector by burst frequency
}
\examples{
a <- acc_example()

freqs(a)

filter_freq(a, 2.5)
}
