% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gulls}
\alias{gulls}
\title{Lesser black-backed gulls}
\format{
A \code{move2} with 5 tracks and 1499 features lasting from
2021-03-03 00:00:00 UTC to 2021-03-04 00:00:00 UTC.

See the \href{https://www.movebank.org/cms/movebank-content/movebank-attribute-dictionary}{Movebank Attribute Dictionary}
for details on data attributes.
}
\source{
\url{https://www.movebank.org/cms/webapp?gwt_fragment=page=studies,path=study985143423}
}
\usage{
gulls
}
\description{
GPS and acceleration data for lesser black-backed gulls
(Larus fuscus, Laridae) breeding at the southern North Sea coast
in Belgium and the Netherlands. Published by
the Research Institute for Nature and Forest (INBO). Data collected by the
\href{(http://lifewatch.be/en/gps-tracking-network-large-birds)}{LifeWatch} GPS
tracking network for large birds for the project/study LBBG_ZEEBRUGGE, using
trackers developed by the University of Amsterdam Bird Tracking System
(\href{http://www.uva-bits.nl}{UvA-BiTS}).
}
\keyword{datasets}
