% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gyro.R
\name{gyro}
\alias{gyro}
\title{Create a \code{gyro} vector}
\usage{
gyro(
  bursts = list(),
  frequency = units::set_units(double(), "Hz"),
  start = NULL
)
}
\arguments{
\item{bursts}{a list of matrices}

\item{frequency}{The frequency of the gyroscope recordings. Either the
same length of \code{bursts} or it will be recycled. If no units are specified,
the frequency is assumed to be in Hz.}

\item{start}{Start time of the burst, in POSIXct format}
}
\description{
Create a \code{gyro} vector
}
