% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colsets.R
\name{gyro_colset}
\alias{gyro_colset}
\title{Specify gyroscope data columns}
\usage{
gyro_colset(
  x = NULL,
  y = NULL,
  z = NULL,
  bursts = NULL,
  axes = NULL,
  frequency = NULL
)
}
\arguments{
\item{x, y, z}{In long-format data, the column name(s) for the
X, Y, and/or Z gyroscope axes.}

\item{bursts}{For burst-format data, the column name containing the raw burst
strings.}

\item{axes}{For burst-format data, the column name containing the axis
labels.}

\item{frequency}{For burst-format data, the column name containing the
sampling frequency.}
}
\value{
A \code{gyro_colset} object.
}
\description{
Define which columns in a \code{move2} object contain gyroscope data.
Use this function to manually specify the columns that should be used when
constructing a \code{gyro} vector with \code{as_gyro()}.

For long-format data (one measurement per row), specify axis columns
with \code{x}, \code{y}, and/or \code{z}.

For burst-format data (where each row contains burst data in a single string),
specify the columns containing the raw burst data, axes, and sampling
frequency.
}
\examples{
# Specify the column names for long-format data for each axis
gyro_colset(x = "my_gyro_x", y = "my_gyro_y", z = "my_gyro_z")

# Long format data may consist of a subset of axes
gyro_colset(x = "my_gyro_x", y = "my_gyro_y")

# Specify the column names for the bursts, axes, and frequency for
# burst-format data
gyro_colset(
  bursts = "my_raw_gyro",
  axes = "my_axes",
  frequency = "my_freq"
)
}
\seealso{
\code{\link[=active_gyro_colsets]{active_gyro_colsets()}} for automatic colset detection, \code{\link[=as_gyro]{as_gyro()}}
to construct a \code{gyro} vector using a colset specification.
}
