% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mag.R
\name{mag}
\alias{mag}
\title{Create a \code{mag} vector}
\usage{
mag(
  bursts = list(),
  frequency = units::set_units(double(), "Hz"),
  start = NULL
)
}
\arguments{
\item{bursts}{a list of matrices}

\item{frequency}{The sampling frequency of the recordings in \code{bursts}. Either the
same length of \code{bursts} or it will be recycled. If no units are specified,
the frequency is assumed to be in Hz.}

\item{start}{Start time of the burst, in POSIXct format}
}
\description{
Create a \code{mag} vector
}
