% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colsets.R
\name{mag_colset}
\alias{mag_colset}
\title{Specify magnetometer data columns}
\usage{
mag_colset(
  x = NULL,
  y = NULL,
  z = NULL,
  bursts = NULL,
  axes = NULL,
  frequency = NULL
)
}
\arguments{
\item{x, y, z}{In long-format data, the column name(s) for the
X, Y, and/or Z magnetometer axes.}

\item{bursts}{For burst-format data, the column name containing the raw burst
strings.}

\item{axes}{For burst-format data, the column name containing the axis
labels.}

\item{frequency}{For burst-format data, the column name containing the
sampling frequency.}
}
\value{
A \code{mag_colset} object.
}
\description{
Define which columns in a \code{move2} object contain magnetometer data.
Use this function to manually specify the columns that should be used when
constructing a \code{mag} vector with \code{as_mag()}.

For long-format data (one measurement per row), specify axis columns
with \code{x}, \code{y}, and/or \code{z}.

For burst-format data (where each row contains burst data in a single string),
specify the columns containing the raw burst data, axes, and sampling
frequency.
}
\examples{
# Specify the column names for long-format data for each axis
mag_colset(x = "my_mag_x", y = "my_mag_y", z = "my_mag_z")

# Long format data may consist of a subset of axes
mag_colset(x = "my_mag_x", y = "my_mag_y")

# Specify the column names for the bursts, axes, and frequency for
# burst-format data
mag_colset(
  bursts = "my_raw_mag",
  axes = "my_axes",
  frequency = "my_freq"
)
}
\seealso{
\code{\link[=active_mag_colsets]{active_mag_colsets()}} for automatic colset detection, \code{\link[=as_mag]{as_mag()}} to
construct a \code{mag} vector using a colset specification.
}
