% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_bursts.R
\name{map_bursts}
\alias{map_bursts}
\title{Apply a function to each element of an \code{acc} or \code{mag} vector}
\usage{
map_bursts(x, .f, simplify = FALSE, .progress = FALSE)
}
\arguments{
\item{x}{An \code{acc} or \code{mag} vector}

\item{.f}{A function to be applied to each element of \code{x}. This can be
supplied in one of the following ways:
\itemize{
\item A named function
\item An anonymous function (e.g., \code{function(.br) nrow(.br) / .fq})
\item A formula (e.g., \code{~ nrow(.br) / .fq})
}

In all cases, use \code{.br} to refer to the burst matrix of each element, \code{.fq}
to refer to the frequency of each element, and \code{.st} to refer to the
start time of each element. See examples.}

\item{simplify}{Logical. If \code{TRUE}, attempts to simplify the output to a
vector. Otherwise, the output will be a list. If the output cannot be
simplified while maintaining a one-to-one correspondence with the input,
an error will be thrown.}

\item{.progress}{Whether to show a progress bar. Use \code{TRUE} to turn on a
basic progress bar, use a string to give it a name.}
}
\description{
This function provides a general framework to apply an arbitrary function
to each element of an \code{acc} or \code{mag} vector while providing access to each
element's burst, frequency, and start time metadata.

Note that some \link[=explore-functions]{common operations} have already been implemented as stand-alone
functions.
}
\details{
This function behaves similarly to the \code{\link[purrr:map]{purrr::map()}} family of functions.
However, \code{map_bursts()} only matches arguments by name, not position. Thus,
the input to \code{.f} must use the specified terminology (\code{.br}, \code{.fq},
and/or \code{.st}) to access specific data from each element.
For a given vector \code{x}:
\itemize{
\item \code{.br} accesses each element of the list returned by \code{\link[=bursts]{bursts()}}
\item \code{.fq} accesses each element of the vector returned by \code{\link[=freqs]{freqs()}}
\item \code{.st} accesses each element of the vector returned by \code{\link[=starts]{starts()}}
}
}
\examples{
a <- acc_example()

# Use `.br` to access the burst matrix for each element:
n_samp <- map_bursts(a, function(.br) nrow(.br))

n_samp

# Use `.fq` to access the frequency value for each element:
burst_len <- map_bursts(a, function(.br, .fq) nrow(.br) / .fq)

burst_len

# Use `.st` to access the start time for each element:
burst_end <- map_bursts(
  a, 
  function(.br, .fq, .st) as.numeric(nrow(.br) / .fq) + .st
)

burst_end

# You can also provide a separately defined function
get_burst_end <- function(.br, .fq, .st, offset = 2) {
  as.numeric(nrow(.br) / .fq) + .st + offset
}

map_bursts(a, get_burst_end)

map_bursts(a, function(.br, .fq, .st) get_burst_end(.br, .fq, .st, offset = 5))

# Use simplify to reduce to a vector format:
map_bursts(a, get_burst_end, simplify = TRUE)

# Note that this will fail if the result cannot be simplified to the same
# length as the input vector
try(
  map_bursts(a, function(.br) .br, simplify = TRUE)
)
}
