% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_merge.R
\name{merge_imu}
\alias{merge_imu}
\title{Merge adjacent bursts in an IMU vector}
\usage{
merge_imu(x, ids = NULL, drop = FALSE)
}
\arguments{
\item{x}{An IMU vector (e.g. \code{acc}, \code{mag}, \code{gyro}).}

\item{ids}{Vector indicating groups to which the elements in \code{x} belong.
If provided, bursts in \code{x} will not be merged across different values of
this vector, even if their timestamps and frequencies align.}

\item{drop}{Logical indicating whether to drop entries that have been merged
into other bursts. If \code{drop = FALSE} (default), the output will have the
same length as the input \code{x}, with \code{NA} values at positions where bursts
were merged into a preceding burst. This is useful for retaining index
matching between the input and output vectors.}
}
\value{
A vector of the same class as \code{x}.
}
\description{
For a given IMU vector, identify temporally adjacent bursts and
merge them into a single burst. Bursts that end at the same time as the
start time of the next burst are considered adjacent. Bursts with different
frequencies or axes will not be merged.
}
\examples{
a <- acc(
  list(cbind(X = 1:60, Y = 1:60), cbind(X = 61:100, Y = 61:100), cbind(X = 101:140)),
  frequency = units::set_units(20, "Hz"),
  start = as.POSIXct(c(0, 3, 5), tz = "UTC")
)

merge_imu(a)
}
