% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/move2Acc-package.R
\docType{package}
\name{move2Acc-package}
\alias{move2Acc}
\alias{move2Acc-package}
\title{move2Acc: Handling bursted acceleration data}
\description{
This package is developed to handel bursted acceleration data. It allows for quick visualizations and convient manipulation.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/bart1/move2Acc}
  \item \url{https://bart1.github.io/move2Acc/}
  \item Report bugs at \url{https://github.com/bart1/move2Acc/issues}
}

}
\author{
\strong{Maintainer}: Bart Kranstauber \email{b.kranstauber@uva.nl} (\href{https://orcid.org/0000-0001-8303-780X}{ORCID})

}
\keyword{internal}
