% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colsets.R
\name{movebank_colsets}
\alias{movebank_colsets}
\alias{movebank_acc_colsets}
\alias{movebank_mag_colsets}
\alias{movebank_gyro_colsets}
\title{View standard Movebank IMU data column sets}
\usage{
movebank_acc_colsets()

movebank_mag_colsets()

movebank_gyro_colsets()
}
\value{
A named list of \code{imu_colset} objects.
}
\description{
Movebank has several standard ways to store data for each IMU sensor. These
functions show the recognized columns for each sensor that can be extracted
from a \code{move2} object by default.
\itemize{
\item \code{movebank_acc_colsets()} — standard column sets for \code{\link[=as_acc]{as_acc()}}.
\item \code{movebank_mag_colsets()} — standard column sets for \code{\link[=as_mag]{as_mag()}}.
\item \code{movebank_gyro_colsets()} — standard column sets for \code{\link[=as_gyro]{as_gyro()}}.
}

To extract IMU data from a \code{move2} with column names that don't correspond to
Movebank's conventions, provide a custom set of IMU columns with
\code{\link[=imu_colset]{imu_colset()}}.
}
\details{
\code{move2} objects store IMU data in two ways:
\itemize{
\item \strong{Long-format} columns store one measurement (possibly for multiple axes)
in a single row.
\item \strong{Burst-format} columns store a burst of measurements as a space-delimited
string. This string must be segmented into axis-specific measurements using
an associated column that indicates the axes present for the bursted data.
A further column provides the sampling frequency of the burst. All three
of these columns must be present to form a valid burst-format column set.
}
}
\examples{
movebank_acc_colsets()
movebank_mag_colsets()
movebank_gyro_colsets()
}
\seealso{
\code{\link[=active_acc_colsets]{active_acc_colsets()}}, \code{\link[=active_mag_colsets]{active_mag_colsets()}}, \code{\link[=active_gyro_colsets]{active_gyro_colsets()}}
to identify column sets present in a given \code{move2} object.
}
