% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_join_closest.R
\name{mt_join_closest}
\alias{mt_join_closest}
\title{Join values onto a move2 object by closest timestamp and ID}
\usage{
mt_join_closest(
  x,
  y,
  ...,
  method = "closest_before",
  track_id_column_y = NULL,
  time_column_y = NULL,
  suffix = c("", ".y")
)
}
\arguments{
\item{x}{A \code{move2} object.}

\item{y}{A data.frame containing columns with track ID and timestamp
information.}

\item{...}{Additional expressions passed to \code{\link[dplyr:join_by]{dplyr::join_by()}}.}

\item{method}{How to match timestamps during the join. One of:
\itemize{
\item \code{"closest_before"} (default): for each record in \code{x}, match to the record in
\code{y} whose timestamp is closest to but not later than the given timestamp in \code{x}.
\item \code{"closest_after"}: for each record in \code{x}, match to the record in \code{y} whose
timestamp is closest to but not earlier than the given timestamp in \code{x}.
\item \code{"exact"}: match records by exact timestamp.
}}

\item{track_id_column_y}{Name of the column containing the track IDs in \code{y}.
Defaults to the \code{\link[move2:column_name]{move2::mt_track_id_column()}} of \code{x}.}

\item{time_column_y}{Name of the column containing the timestamps in \code{y}.
Defaults to the \code{\link[move2:column_name]{move2::mt_time_column()}} of \code{x}.}

\item{suffix}{If there are non-joined duplicate variables in x and y, these
suffixes will be added to the output to disambiguate them. Should be a
character vector of length 2.

Note that if the first element of the suffix
is not \code{""}, the timestamp column name of \code{x} may be changed during the
join, in which case an \code{sf} object will be returned, not a \code{move2} object.}
}
\value{
\code{x} with additional columns from \code{y} attached
}
\description{
Joins a \code{move2} object with a data.frame, matching on track ID and timestamp.
Values can be matched with the closest timestamp before each given
timestamp in the input \code{move2}, the closest timestamp after each given
timestamp in the input \code{move2}, or by an exact match (a standard join).

The join will always use the \code{\link[move2:column_name]{move2::mt_track_id_column()}} and
\code{\link[move2:column_name]{move2::mt_time_column()}} of the input \code{move2} as the track ID and timestamp,
respectively. To join on different columns, set them beforehand with
\code{\link[move2:column_name]{move2::mt_set_track_id_column()}} and/or \code{\link[move2:column_name]{move2::mt_set_time_column()}}.
}
\examples{
\dontrun{
library(move2)
library(dplyr)

alb <- albatrosses()

vedbas <- alb |> 
  mutate(v = vedba(as_acc(alb, drop = FALSE))) |> 
  select(v) |> 
  sf::st_drop_geometry() |> 
  filter(!is.na(v))

alb_gps <- alb[!sf::st_is_empty(alb), ]

# Match based on closest timestamp in y prior to each timestamp in x
alb_joined <- mt_join_closest(alb_gps, vedbas)

alb_joined |> select(v, timestamp.y)

# Match based on closest timestamp in y following each timestamp in x
alb_joined <- mt_join_closest(alb_gps, vedbas, method = "closest_after")

alb_joined |> select(v, timestamp.y)

# Provide additional conditions to match records on, for instance to match
# closest prior timestamp only for GPS records (with sensor type 653):
vedbas$sensor <- 653
alb$sensor_type_id <- as.numeric(alb$sensor_type_id)

mt_join_closest(alb, vedbas, sensor_type_id == sensor) |>
  select(v, timestamp.y, sensor_type_id)
}
}
