% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak_frequency.R
\name{peak_frequency}
\alias{peak_frequency}
\title{Calculate the peak frequency per axis for acceleration bursts}
\usage{
peak_frequency(x, resolution = NA)
}
\arguments{
\item{x}{An \code{acc} vector}

\item{resolution}{A scalar with the \link[units:units]{units} Hertz}
}
\value{
returns a list with the same length as \code{x} with the peak frequency per axis
}
\description{
Calculate the peak frequency per axis for acceleration bursts
}
\details{
To increase the resolution of the result zero padding can be used. This can be controlled using the resolution
argument. Note that increasing resolution without increasing the number of samples in a acceleration burst has a
limited ability to get closer to the true frequency.
}
\examples{
a <- acc(
  list(
    cbind(
      X = sin(1:200 / (5  / (pi * 2))),
      Z = cos(1:200 / (80 / (pi * 2)))
    )
  ), 
  units::set_units(400, "Hz")
)

peak_frequency(a)

peak_frequency(a, units::set_units(.25, "Hz"))

# Increasing resolution more
peak_frequency(a, units::set_units(.005, "Hz"))

a <- acc(
  list(
    cbind(
      X = sin((1:200) / (5 / (pi * 2))),
      Z = cos(80 + 1:200 / (80 / (pi * 2)))
    )
  ), 
  units::set_units(400, "Hz")
)

peak_frequency(a, units::set_units(.005, "Hz"))
}
