% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_imu.R
\name{plot.imu_summary}
\alias{plot.imu_summary}
\title{Plot an \code{imu_summary}}
\usage{
\method{plot}{imu_summary}(x, ...)
}
\arguments{
\item{x}{An \code{imu_summary} object (returned by \code{\link[=summary.imu]{summary.imu()}}).}

\item{...}{Passed to \code{\link[graphics:hist]{graphics::hist()}}.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Produces a multi-panel histogram of the distributions stored in an
\code{imu_summary} object.
}
\examples{
s <- summary(acc_example())
plot(s)
}
