% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_time.R
\name{plot_time}
\alias{plot_time}
\title{Plot bursts over time}
\usage{
plot_time(x, time, ylab = "Value")
}
\arguments{
\item{x}{An IMU vector (e.g. \code{acc}, \code{mag}, \code{gyro}).}

\item{time}{a \code{POSIXct} with the start time of bursts}

\item{ylab}{A character with the y axis label}
}
\description{
Plot bursts over time
}
