% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/continuous_acc.R
\name{split_continuous_acc}
\alias{split_continuous_acc}
\title{Split an \code{acc} object at regular intervals}
\usage{
split_continuous_acc(x, interval)
}
\arguments{
\item{x}{An \code{acc} vector}

\item{interval}{Numeric or units object defining the time intervals at which
\code{x} will be split. If no units are provided, the interval is assumed to
be in period units of \code{x} (i.e., 1 divided by the frequency units).}
}
\value{
An \code{acc} vector
}
\description{
Split the bursts in an \code{acc} object into bursts of a given time duration.
}
\examples{
a <- acc(
  c(acc_burst_example(1:60, 1:60), acc_burst_example(101:140)),
  frequency = c(units::set_units(20, "Hz"), units::set_units(40, "Hz")),
  start = as.POSIXct(c(0, 10), tz = "UTC")
)

x <- split_continuous_acc(a, units::set_units(1, "s"))
x

# Start times handled automatically 
starts(x)

# Records are not guaranteed to have same duration depending on interval
# and burst size:
x <- split_continuous_acc(a, 0.7)

burst_dur(x)
}
