% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colsets.R
\name{valid_gyro_colsets}
\alias{valid_gyro_colsets}
\alias{gyro_colset_burst}
\alias{gyro_colset_xyz}
\title{Valid gyroscope data column sets}
\usage{
valid_gyro_colsets()

gyro_colset_burst()

gyro_colset_xyz()
}
\value{
For \code{valid_gyro_colsets()}, a list of \code{gyro_colset} objects
containing valid column sets. Otherwise, a \code{gyro_colset} object.
}
\description{
These sets of columns can be used by \code{\link[=as_gyro]{as_gyro()}} when parsing gyroscope
bursts contained in a \code{move2} object. A \code{move2} object must contain one
of these column sets to be processed by \code{as_gyro()}.
\itemize{
\item \code{gyro_colset_burst()} must be present in its entirety within a data
source to be used when parsing gyroscope data.
\item For \code{gyro_colset_xyz()}, any subset of the set's columns can be used to
parse gyroscope data.
}

To determine the default columns that will be used by \code{as_gyro()} for a given
\code{move2} object, see \code{\link[=active_gyro_colsets]{active_gyro_colsets()}}.
}
\examples{
valid_gyro_colsets()
}
