% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colsets.R
\name{active_acc_colsets}
\alias{active_acc_colsets}
\title{Identify acceleration columns present in a \code{move2} object}
\usage{
active_acc_colsets(x)
}
\arguments{
\item{x}{A \code{move2} containing acceleration data as collected by EOBS,
Ornitela, or similar tracking devices. Most of the time this will be
either loaded from disk using \link[move2:mt_read]{move2::mt_read} or downloaded using
\link[move2:movebank_download_study]{move2::movebank_download_study}.}
}
\value{
A list of \code{acc_colset} objects.
}
\description{
\code{active_acc_colsets()} determines the sets of columns that will be used
by default to construct acceleration bursts with \code{as_acc()}. Column sets
are processed independently, but a given \code{move2} may contain multiple
column sets with acceleration data.
}
\details{
\code{move2} objects store acceleration data in two ways: long-format
acceleration columns and burst-format acceleration columns.

Long-format columns store one acceleration measurement (possibly for multiple
axes) in a single row. Note that
not all axes need to be present for a long-format column set to be considered
active in a \code{move2} object.

Burst-format columns store a burst of acceleration data as a space-delimited
string. This string must be segmented into axis-specific measurements using
an associated column that indicates the axes present for the bursted data.
A further column provides the sampling frequency of the burst.
All three of these columns must be present for a burst-format column set
to be considered active in a \code{move2} object.

Standard long-format column sets for data from
Movebank include \code{\link[=acc_colset_raw_xyz]{acc_colset_raw_xyz()}} and \code{\link[=acc_colset_xyz]{acc_colset_xyz()}}. Standard
burst-format column sets for data from Movebank include \code{\link[=acc_colset_eobs]{acc_colset_eobs()}}
and \code{\link[=acc_colset_burst]{acc_colset_burst()}}.

If your input data use different column names for these columns, use
\code{\link[=acc_colset]{acc_colset()}} to specify the column names that correspond to each of the
axes (for long-format data) or burst data and associated metadata (for
burst-format data).
}
\examples{
active_acc_colsets(albatrosses())

# Multiple colsets may be available
active_acc_colsets(move2::mt_stack(albatrosses(), gulls()))

# Missing long-format axes are not included in the set
g <- gulls()
g$acceleration_raw_x <- NULL
active_acc_colsets(g)

# Columns with no data are also removed
g$acceleration_raw_y <- NA
active_acc_colsets(g)

# Some column sets must be present in their entirety
alb <- albatrosses()
alb$eobs_acceleration_axes <- NULL

\dontrun{
  active_acc_colsets(alb)
}
}
\seealso{
\code{\link[=valid_acc_colsets]{valid_acc_colsets()}} for currently supported default colsets,
\code{\link[=as_acc]{as_acc()}} to build an \code{acc} vector from a \code{move2} object.
}
