% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acc_calibrate.R
\name{acc_calibrate}
\alias{acc_calibrate}
\title{Convert raw acceleration values to physical units in an \code{acc} vector}
\usage{
acc_calibrate(acc, calibration)
}
\arguments{
\item{acc}{An \code{acc} vector.}

\item{calibration}{An \code{acc_calibration} object containing the calibration
function(s) to apply to each burst in \code{acc}. See \code{\link[=acc_calibration]{acc_calibration()}} to
specify calibration functions.}
}
\value{
An \code{acc} vector of the same length as the input with calibrated
burst matrices.
}
\description{
Applies calibration function(s) to bursts of raw acceleration ADC values
in an \code{acc} vector. Specify a set of calibration functions
with \code{\link[=acc_calibration]{acc_calibration()}}.
}
\examples{
a <- acc_example()

acc_calibrate(a, acc_calibration("ornitela"))

acc_calibrate(a, acc_calibration("eobs", tag_id = 1000))

acc_calibrate(a, acc_calibration(offset = 2048, slope = 0.001))

# Specify different calibration parameters for each burst.
# Calibrations will be mapped to the input `acc` by index.
acc_calibrate(
  a, 
  acc_calibration(offset = c(2048, 2046), slope = c(0.001, 0.002))
)
}
\seealso{
\code{\link[=acc_calibration]{acc_calibration()}} and \code{\link[=as_acc_calibration]{as_acc_calibration()}} to
construct calibrations for use with \code{acc_calibrate()}.
}
