% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_acc.R
\name{as_acc}
\alias{as_acc}
\alias{as_acc.default}
\alias{as_acc.move2}
\title{Convert to acc}
\usage{
as_acc(x, ...)

\method{as_acc}{default}(x, ...)

\method{as_acc}{move2}(
  x,
  acc_cols = NULL,
  min_frq = 1,
  merge_continuous = TRUE,
  drop = TRUE,
  ...
)
}
\arguments{
\item{x}{A \code{move2} containing acceleration data as collected by EOBS,
Ornitela, or similar tracking devices. Most of the time this will be
either loaded from disk using \link[move2:mt_read]{move2::mt_read} or downloaded using
\link[move2:movebank_download_study]{move2::movebank_download_study}.}

\item{...}{currently not used}

\item{acc_cols}{Character vector of column names identifying the columns in
\code{x} that contain the acceleration data to be used when constructing the
output \code{acc} vector. By default, uses the columns returned by
\code{\link[=active_acc_cols]{active_acc_cols()}}. See \code{\link[=valid_acc_colsets]{valid_acc_colsets()}} to identify supported
acceleration column names.}

\item{min_frq}{Numeric value indicating the
minimum allowable within-burst data collection frequency when identifying
bursts in long-format acceleration data. Any two adjacent timestamps
that fall outside of the period defined by this frequency will be split
into separate bursts. If no units are provided, this value is assumed to
be in Hz.

Ignored if data are already in predefined bursts (e.g. e-obs).}

\item{merge_continuous}{Logical value indicating whether to merge
adjacent acceleration bursts. Two adjacent bursts can be merged if the
first burst ends at the same time that the second starts and the
burst frequency is identical between the two. This is useful for
processing continuous acceleration data that have been stored in chunks
split at regular intervals.}

\item{drop}{Logical indicating whether empty acceleration objects should
be dropped from the output. If \code{drop = FALSE}, then the length of the
output will match the number of rows in the input data \code{x} and acceleration
bursts will be stored at the index location corresponding to the start time
of the burst.}
}
\description{
In many cases the \code{as_acc} function will directly create an acceleration
vector from input data
}
\details{
The resulting vector will be as long as the input. This means it
can, for example, be added as a column to a \code{data.frame}. For some tags
this means \code{NA} values are inserted when one burst is stored over multiple
rows of a \code{data.frame}.
}
