% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acc.R, R/funs.R, R/mag.R
\name{is_acc}
\alias{is_acc}
\alias{n_axis}
\alias{explore-functions}
\alias{is_uniform}
\alias{bursts}
\alias{bursts<-}
\alias{freqs}
\alias{freqs<-}
\alias{starts}
\alias{starts<-}
\alias{burst_dur}
\alias{n_samples}
\alias{burst_units}
\alias{is_mag}
\title{Functions to explore an \code{acc} vector}
\usage{
is_acc(x)

n_axis(x)

is_uniform(x)

bursts(x)

bursts(x) <- value

freqs(x)

freqs(x) <- value

starts(x)

starts(x) <- value

burst_dur(x)

n_samples(x)

burst_units(x)

is_mag(x)
}
\arguments{
\item{x}{an acc vector}

\item{value}{Replacement value.}
}
\description{
Functions to explore an \code{acc} vector
}
\examples{
x <- acc(
  bursts = list(
    cbind(X = sin(1:30 / 10), Y = cos(1:30 / 10), Z = 1),
    cbind(X = sin(1:20 / 10 + 2), Y = cos(1:20 / 10 + 3))
  ),
  frequency = units::as_units(c(20, 30), "Hz")
)
x <- c(x, NA)
n_axis(x)
n_samples(x)
is_uniform(x)
length(x)
is.na(x)
na.omit(x)
 y <- acc(
  bursts = list(
    cbind(X = sin(1:20 / 10), Y = cos(1:20 / 10)),
    cbind(X = sin(1:20 / 10 + 2), Y = cos(1:20 / 10 + 3))
  ),
  frequency = units::as_units(c(20, 20), "Hz")
)
is_uniform(y)
}
