% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_time.R
\name{plot_time}
\alias{plot_time}
\title{Plot bursts over time}
\usage{
plot_time(x, ylab = "Value")
}
\arguments{
\item{x}{An IMU vector (e.g. \code{acc}, \code{mag}, \code{gyro}).}

\item{ylab}{A character with the y axis label}
}
\description{
Plot the trace of IMU values from an IMU vector with time on the x-axis.
}
\details{
If the bursts in the input come from multiple sources, traces may be
combined incorrectly. See examples.
}
\examples{
\dontshow{if (rlang::is_installed(c("dygraphs", "move2"))) withAutoprint(\{ # examplesIf}
plot_time(acc_example())

# If bursts come from multiple sources (in this case, deployments),
# then lines from different bursts may be incorrectly connected:
alb <- albatrosses()
a <- as_acc(alb, drop = FALSE)

plot_time(a)

# To avoid this issue, plot only a single deployment's values:
plot_time(a[move2::mt_track_id(alb) == "4261-2228"])
\dontshow{\}) # examplesIf}
}
