% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/continuous_acc.R
\name{split_continuous_acc}
\alias{split_continuous_acc}
\title{Split an \code{acc} object at regular intervals}
\usage{
split_continuous_acc(x, interval)
}
\arguments{
\item{x}{An \code{acc} vector}

\item{interval}{Numeric or units object defining the time intervals at which
\code{x} will be split. If no units are provided, the interval is assumed to
be in period units of \code{x} (i.e., 1 divided by the frequency units).}
}
\value{
A list of \code{acc} vectors, the same length as \code{x}. Each element
contains the split pieces of the corresponding input burst.
}
\description{
Split the bursts in an \code{acc} object into bursts of a given time duration.
The result is a list of \code{acc} vectors of the same length as the input.
}
\examples{
a <- acc(
  c(acc_burst_example(1:60, 1:60), acc_burst_example(101:140)),
  frequency = c(units::set_units(20, "Hz"), units::set_units(40, "Hz")),
  start = as.POSIXct(c(0, 10), tz = "UTC")
)

x <- split_continuous_acc(a, units::set_units(1, "s"))
x

# Flatten to a single acc vector
flat <- purrr::reduce(x, c)
flat

# Start times are updated to match the start of each split component
starts(flat)

# Use merge_continuous_acc() on flat
identical(merge_continuous_acc(flat), a)

\dontrun{
# In a dataframe, split and unnest to retain index matching
library(dplyr)
library(tidyr)

tbl <- tibble::tibble(id = c("a", "b"), burst = a)

tbl <- tbl |>
  mutate(burst = split_continuous_acc(burst, units::set_units(1, "s"))) |>
  unnest(burst) |>
  mutate(timestamp = starts(burst))
  
tbl
  
# Use merge_continuous_acc() to recover original bursts
tbl |> 
  mutate(burst = merge_continuous_acc(burst, acc_ids = id, drop = FALSE))
}
}
