% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colsets.R
\name{valid_acc_colsets}
\alias{valid_acc_colsets}
\title{List valid acceleration data column sets}
\usage{
valid_acc_colsets()
}
\value{
List of vectors of valid column sets
}
\description{
These columns are used by \code{as_acc()} when parsing acceleration bursts
contained in a \code{move2} object. A \code{move2} object must contain one
of these column sets to be processed by \code{as_acc()}.

If the \code{move2} object contains one of the first two sets, it must contain
all three of the listed columns in that set. If it contains one of the
final three sets, any subset of the set's listed columns are sufficient.

To determine which columns will be used by \code{as_acc()} for a given
\code{move2} object, see \code{\link[=active_acc_cols]{active_acc_cols()}}.
}
\examples{
valid_acc_colsets()
}
