% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colsets.R
\name{valid_acc_colsets}
\alias{valid_acc_colsets}
\alias{acc_eobs_cols}
\alias{acc_burst_cols}
\alias{acc_xyz_cols}
\alias{acc_raw_xyz_cols}
\title{Valid acceleration data column sets}
\usage{
valid_acc_colsets()

acc_eobs_cols()

acc_burst_cols()

acc_xyz_cols()

acc_raw_xyz_cols()
}
\value{
For \code{valid_acc_colsets()}, a list of vectors of valid column sets.
Otherwise, a character vector containing the names of the
}
\description{
These sets of columns can be used by \code{\link[=as_acc]{as_acc()}} when parsing acceleration
bursts contained in a \code{move2} object. A \code{move2} object must contain one
of these column sets to be processed by \code{as_acc()}.
\itemize{
\item \code{acc_eobs_cols()} and \code{acc_burst_cols()} must be present in their entirety
within a data source to be used when parsing acceleration data.
\item For \code{acc_xyz_cols()} and \code{acc_raw_xyz_cols()}, any
subset of the set's columns can be used to parse acceleration
data.
}

To determine the default columns that will be used by \code{as_acc()} for a given
\code{move2} object, see \code{\link[=active_acc_colsets]{active_acc_colsets()}}.
}
\examples{
valid_acc_colsets()
}
